chore: split into cli & core modules
